Add package download

+++ src/main/java/ru/tilipod/feign/YandexFeignClient.java
@@ -19,5 +19,5 @@ public interface YandexFeignClient {
     ResponseEntity<YandexResourceResponse> getMetadataForFiles(@RequestHeader String authorization,
                                                                @RequestParam String path,
                                                                @RequestParam Integer limit,
-                                                               @RequestParam Integer offset);
+                                                               @RequestParam Long offset);
 }
